chore: move stack trace related tests to separate module (#1900)
